Brought nAudio in solution.
This is because there is a few changes that need to happen to easily support setting the default device using the Ipolicyconfig interfaces in the MMAPI

I will internalize all of the nAudio source is attempt to abstract the crap that people don't need when interfacing with the library.

Big props to nAudio! Great library