Added native collection handling to Evaluate() for JS.
Removes the need to know the "internal" types of the engine. Obviously there is a required knowledge when trying to pull values out of an executed script.

This only really has advantages while testing, or in custom implementations. Normal execution should not require manipulation of these objects in the Javascript space.