Changed the entire audio device changed event system.
There is now inheritance which will help with identifying the events, and it also allows for more specific event information to be passed in the event object.

Have also changed the way CoreAudioDevice reacts to changed events.

If a CoreAudioDevice detects a stage change it should also update the internal audio properties. Fixes #16.