SessionCOntroller now implements IEnmerable for easier linq queries.
CoreAudioSession - first implmentation of the mute function. Needs to be revised to cache the value and then update the value ont he event.

Sessions seem to be working, there's a weird issue with child process and audio sessions not removing themselves correctly. Noticed when using Chrome as  a test case.